restructured translateCategories into an API
