+++ src/components/workoutListing/WorkoutListing.jsx
@@ -1,0 +1,5 @@
+
+const WorkoutListing = (props) => {
+};
+
+export default WorkoutListing;
