+++ src/components/workoutListing/workoutItem/WorkoutItem.jsx
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+const StyledLi = styled.li`
+  padding: 15px 18px 10px 14px;
+`;
+
+const WorkoutItem = (props) => {
+  return (
+    <StyledLi>
+
+    </StyledLi>
+  )
+};
+
+export default WorkoutItem;
